CPPS-1: Give up on CompileTimeString but preserve in separate file for later
